\name{exampleBLData}
\alias{exampleBLData}
\docType{data}
\title{beadLevelData object from an example experiment}
\description{

The data in this package are a subset of the MAQC bead-level data available in the beadarrayUseCases package. Bead-level refers to the availability of intensity and location information for each bead on each BeadArray in an experiment.  In this dataset, BeadArrays were hybridized with either Universal Human Reference RNA (UHRR, Stratagene) or Brain Reference RNA (Ambion) as used in the MAQC project. This object is a representation of the bead-level data for 2 arrays and was created by the beadarray package.


}
\usage{data(exampleBLdata)}
\seealso{\code{\link[beadarray:class-beadLevelData]{beadLevelData}}}
\keyword{datasets}
