\name{exampleSummaryData}
\alias{exampleSummaryData}
\docType{data}
\title{ExpressionSetIllumina object for the example experiment}
\description{
  \code{exampleSummaryData} is an object of class \code{ExpressionSetIllumina} which is a summarized version of the bead-level data  distributed with the \code{BeadArrayUseCases} package. There are 12 MAQC samples in the object which are either Brain or UHRR. Unlogged and log2 versions of the data are provided. See the vignette for \code{beadarrayExampleData} for details of how these data were created.
}
\usage{data(exampleSummaryData)}
\keyword{datasets}
